==> 2. Scripts/1. Importing libraries and Python data types.ipynb <==
# 01. Importing libraries

# Import libraries
import pandas as pd
import numpy as np
import os

# 02. Data types

# Integers
a = 85
b = 15
c = 101
d = 1
e = 84
f = 0
a + b, c-d, c - a + e + f

# Floating-point numbers
g = 19.0819
h = 0.236
g / h, h / g

# String word
i = 'jup'
j = 'yter'
i + j

# String sentence
k = 'When we strive to be'
l = 'come better than we are, '
m = 'everything around us '
n = 'becomes better'
o = ' too.'
k + l + m + n + o

# Bonus
p = 'fun'
p * 100
