==> 2. Scripts/3. Data Wrangling & Subsetting.ipynb <==
# 01. Importing libraries

#Import libraries
import pandas as pd
import numpy as np
import os

# 02. Importing data

#Create a path
path = r'C:\Users\opetr\Documents\Instacart Basket Analysis'

#Import orders.csv
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), index_col = False)

#Import products.csv
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

# 03. Data wrangling

## Dropping columns

#Drop column eval_set from dataframe
df_ords.drop(columns = ['eval_set'])

#Permanently overwriting the dataframe to remove eval_set column
df_ords = df_ords.drop(columns = ['eval_set'])

## Missing values

#Find missing values in a column
df_ords['days_since_prior_order'].value_counts(dropna = False)

## Frequency count

#Exclude the missing values to find existing values only in a column
df_ords['days_since_prior_order'].value_counts()

## Renaming columns

#Change the name of column order_dow to orders_day_of_week
df_ords.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)

## Changing the data type

#Change data type to string in a column
df_ords['order_id'] = df_ords['order_id'].astype('str')

#Verify that change was executed. O stands for object
df_ords['order_id'].dtype

## Transposing data

#Import departments.csv
df_dep = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'departments.csv'), index_col = False)

#Glance at new dataframe
df_dep.head()

#Transpose from a wide format to a long format
df_dep_t = df_dep.T

#Adding an index to the dataframe
df_dep_t.reset_index()

#Create an object that contains the first row of the dataframe
new_header = df_dep_t.iloc[0]

#Check the contants of row 0
new_header

#Create a new dataframe to copy everything from the first row onward
df_dep_t_new = df_dep_t[1:]

#Check on the new dataframe
df_dep_t_new

#Set newly created variable as the new header
df_dep_t_new.columns = new_header

#Check on the new header
df_dep_t_new

## Data dictionary

#Turn dataframe in a dictionary
data_dict = df_dep_t_new.to_dict('index')

#Check on newly created data dictionary
data_dict

#Check first 5 entries in products.csv
df_prods.head()

#Use the data dictionary to determine what an entry in the department_id column means
print(data_dict.get('19'))

## Subsetting

#Create a subset from the df_prods dataframe that only contains data from the snacks department
df_snacks =  df_prods[df_prods['department_id']==19]

#Right side of the command (the indexing code itself)
df_prods['department_id']==19

#Show output in columns within the df_prods dataframe
df_prods[df_prods['department_id']==19]

#Variation using the loc function to achieve the same results
df_snacks_2 = df_prods.loc[df_prods['department_id'] == 19]

#Variation using the loc function to achieve the same results
df_snacks_3 = df_prods.loc[df_prods['department_id'].isin([19])]

# TASK 4.4

#2. Change the data type
df_ords['user_id'].astype('str')

#3. Change column name
df_ords.rename(columns = { 'orders_day_of_week' : 'order_day_of_week'})

#4. Find the frequency. Answer: the busiest hour is 10 AM
df_ords['order_hour_of_day'].value_counts()

#5. Use the data dictionary to find the meaning behind a value of 4 in the department_id column
print(data_dict.get('4'))

#6a. Create a subset containing only breakfast item sales
df_breakfast = df_prods[df_prods['department_id'] == 14]

#6b. Check on results of new subset
df_breakfast.head()

#7a. Create a subset that includes items from the following departments: alcohol, deli, beverages, and meat/seafood
df_dinner_party = df_prods.loc[df_prods['department_id'].isin([5,7,12,20])]

#7b. Check on results of new subset
df_dinner_party.head()

#8. Count rows in the Dinner party dataframe. Answer: 7650 rows
df_dinner_party.shape

#9a. Extract information about user_id 1 by creating a subset
df_cust_1 = df_ords.loc[df_ords['user_id'] == 1]

#9b. Check on results of new subset
df_cust_1.head()

#10. Provide basic states on user_id 1
df_cust_1.describe()

#12. Export df_ords dataframe in the Prepared Data folder
df_ords.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_wrangled.csv'))

#13. Export df_dep_t_new dataframe in the Prepared Data folder
df_dep_t_new.to_csv(os.path.join(path, '02 Data','Prepared Data', 'departments_wrangled.csv'))
