==> 2. Scripts/4. Data Consistency Checks.ipynb <==
# 01. Import libraries

#Import libraries
import pandas as pd
import numpy as np
import os

# 02. Import data

#Import data
path = r'C:\Users\opetr\Documents\Instacart Basket Analysis'

#Import products.csv
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data',  'products.csv'), index_col = False)

#Import orderswrangled.csv
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_wrangled.csv'), index_col = False)

# 03. Data consistency checks

## Mixed-type data

#Create a dataframe
df_test = pd.DataFrame()

#Create a mixed type column
df_test['mix'] = ['a', 'b', 1, True]

#Check on test dataframe
df_test.head()

#Check if df contains mixed-type columns
for col in df_test.columns.tolist():
  weird = (df_test[[col]].applymap(type) != df_test[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_test[weird]) > 0:
    print (col)

#Change data type to string
df_test['mix'] = df_test['mix'].astype('str')

## Missing values

#Finding missing values
df_prods.isnull().sum()

#Create subset to show missing values
df_nan = df_prods[df_prods['product_name'].isnull() == True]

#Check on newly created subset
df_nan

#Check initial number of rows in the df
df_prods.shape

#Create new clean df
df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]

#Check shape after change
df_prods_clean.shape

## Duplicates

#Finding duplicates
df_dups = df_prods_clean[df_prods_clean.duplicated()]

#Check new duplicate df
df_dups

#Check initial number of rows in the df
df_prods_clean.shape

#Create new df that excludes duplicates
df_prods_clean_no_dups = df_prods_clean.drop_duplicates()

#Check shape after change
df_prods_clean_no_dups.shape

## Exporting changes

#Export new df
df_prods_clean_no_dups.to_csv(os.path.join(path, '02 Data', 'Prepared Data', 'products_checked.csv'))

# TASK 4.5

#2. Run the df.describe() function on your df_prods dataframe
df_prods.describe()

#### The max price for a product is 99,999 which could be mistake since the mean is 9.99 and these are grocery store products that do not tend to be this expensive. I would communicate finding to client.

#3. Check for mixed-type data in your df_ords dataframe
for col in df_ords.columns.tolist():
  weird = (df_ords[[col]].applymap(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords[weird]) > 0:
    print (col)

#4. If you find mixed-type data, fix it. The column in question should contain observations of a single data type
#Answer: No mixed-type data found in this dataframe

#5. Run a check for missing values in your df_ords dataframe
df_ords.isnull().sum()

df_ords.user_id.max()

#### There are 206,209 missing values in the days_since_prior_order column. Since there are 206,209 individual User IDs in the user_id column, I hypothesize that that each User ID has an average of 1 entry missing in the days_since_prior_order column (one user_id might have 2 missing, whereas another might have none). It is also possible and more plausible that each User has a value missing due to no prior order from when they were a first time customer

#6. Address the missing values using an appropriate method
#Create a subset for missing values
df_nan = df_ords[df_ords['days_since_prior_order'].isnull() == True]

#Call on new subset and inspect missing values as it relates to user_id
df_nan

#### We can see the user_id goes in an uninterrupted sequence from 1 to 206,209, proving that each customer has one missing value in the column in question. The observations corresponding to each missing value could be useful in analyzing what a first time customer purchases. For this reason, we should not remove them. We should not impute a new value based on the mean or median due to the same reason - this information could be useful farther down the road. We could create a new variable that acts like a flag based on the missing value, but I think the best decision in this case would be to create a value of zero instead, showing that there were no prior purchases made. This would also make these observations easier to call on in the future, and it will be easy for anyone inspecting the data in the future to draw accurate conclusions.

#Update missing values to 0
df_ords['days_since_prior_order'].fillna(0, inplace=True)

#Check that each User ID has no missing values and one value of 0
df_ords.head(30)

#7. Run a check for duplicate values in your df_ords data
df_ords_dups = df_ords[df_ords.duplicated()]

#Check on duplicate subset
df_ords_dups

##### Address the duplicates using an appropriate method. Answer: There are no duplicate values in df_ords

#9a. Export your final, cleaned df_prods and df_ords data as “.csv” files in your “Prepared Data” folder
##df_prods exported before Task began
df_ords.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_cleaned.csv'))
