==> 2. Scripts/5. Combining & Exporting Data.ipynb <==
# 01. Import libraries

#Import libraries
import pandas as pd
import numpy as np
import os

# 02. Import data

#Import data
path = r'C:\Users\opetr\Documents\Instacart Basket Analysis'

#Import wrangled orders data set
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_wrangled.csv'), index_col = False)

# 03. Methods for combining data

### Create dictionaries of data to experiment with

# Define a dictionary containing January 2020 data
data1 = {'customer_id':['6732', '767', '890', '635'],
    'month':['Jan-20', 'Jan-20', 'Jan-20', 'Jan-20'],
    'purchased_meat':[0, 13, 3, 4],
    'purchased_alcohol':[1, 2, 10, 0],
    'purchased_snacks':[10, 5, 1, 7]}

# Define a dictionary containing February 2020 data
data2 = {'customer_id':['6732', '767', '890', '635'],
    'month':['Feb-20', 'Feb-20', 'Feb-20', 'Feb-20'],
    'purchased_meat':[0, 10, 5, 3],
    'purchased_alcohol':[2, 4, 14, 0],
    'purchased_snacks':[15, 3, 2, 6]}

#Convert the dictionary into dataframe
df = pd.DataFrame(data1,index=[0, 1, 2, 3])

#Convert the dictionary into dataframe
df_1 = pd.DataFrame(data2,index=[0, 1, 2, 3])

df

df_1

### Concatenating data

#Create list containing the dfs needed to combine
frames = [df, df_1]

#Concatenate
df_concat = pd.concat(frames)

df_concat

#Create wide-formate df
df_concat = pd.concat(frames, axis = 1)

df_concat

### Appending data

df_appended = df.append(df_1)

df_appended

# Create data with different columns from df
data3 = {'customer_id':['6732', '767', '890', '635'],
    'month':['Jan-20', 'Jan-20', 'Jan-20', 'Jan-20'],
    'days_purchased_on':[0, 13, 3, 4]}

# Convert to dataframe
df_2 = pd.DataFrame(data3,index=[0, 1, 2, 3])

df_append_test = df.append(df_2)

df_append_test

### Merging data

df_merged = df.merge(df_2, on = 'customer_id')

df_merged

df_merged = df.merge(df_2, on = ['customer_id', 'month'])

df_merged

df_merged = df.merge(df_2, on = ['customer_id', 'month'], indicator = True)

df_merged

df_merged['_merge'].value_counts()

# Test merge without overwriting
pd.merge(df,df_2, on = ['customer_id', 'month'], indicator = True)

df.merge(df_2, on = ['customer_id', 'month'], how = 'inner')

# 04. Merging Instacart data

#Import df_ords_prior
df_ords_prior = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'order_products_prior.csv'), index_col = False)

#Check the output
df_ords_prior.head()

df_ords_prior.shape

df_ords.shape

#Merging
df_merged_large = df_ords.merge(df_ords_prior, on = 'order_id', indicator = True)

df_merged_large['_merge'].value_counts()

# 05. Exporting data in Pickle format

#Export data to .pkl
df_merged_large.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_combined.pkl'))
