==> 2. Scripts/6. Deriving New Variables.ipynb <==
# 01. Import libraries

# Import libraries
import pandas as pd
import numpy as np
import os

# 02. Import data

# Create path
path = r'C:\Users\opetr\Documents\Instacart Basket Analysis'

# Import ords_prods_merged
ords_prods_merged = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_merged.pkl'))

# 03. If-Statements with a user-defined function

# Create a subset
df = ords_prods_merged[:1000000]

# Check output
df.shape

# Create a user-defined function
def price_label(row):

  if row['prices'] <= 5:
    return 'Low-range product'
  elif (row['prices'] > 5) and (row['prices'] <= 15):
    return 'Mid-range product'
  elif row['prices'] > 15:
    return 'High range'
  else: return 'Not enough data'

# Apply the function
df['price_range'] = df.apply(price_label, axis=1)

# Check output
df['price_range'].value_counts(dropna = False)

# Check most expensive products in the subset
df['prices'].max()

# 04. If-Statements with with the loc() function

# Create condition 1
ords_prods_merged.loc[ords_prods_merged['prices'] > 15, 'price_range_loc'] = 'High-range product'

# Create condition 2
ords_prods_merged.loc[(ords_prods_merged['prices'] <= 15) & (ords_prods_merged['prices'] > 5), 'price_range_loc'] = 'Mid-range product'

# Create condition 3
ords_prods_merged.loc[ords_prods_merged['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

ords_prods_merged['price_range_loc'].value_counts(dropna = False)

# 05. If-Statements with For-Loops

# Find the frequency of the orders_day_of_the_week column
ords_prods_merged['orders_day_of_week'].value_counts(dropna = False)

# Create a list using a for-loop
result = []

for value in ords_prods_merged["orders_day_of_week"]:
  if value == 0:
    result.append("Busiest day")
  elif value == 4:
    result.append("Least busy")
  else:
    result.append("Regularly busy")

# Check output
result

# Incorporate results in df as a new column
ords_prods_merged['busiest_day'] = result

# Check output
ords_prods_merged['busiest_day'].value_counts(dropna = False)

# TASK 4.7

# 2. Create new columns for the 2 “Busiest days” and the 2 “Slowest days”
## Find the frequency of the orders_day_of_the_week column
ords_prods_merged['orders_day_of_week'].value_counts(dropna = False)

# Create a list using a for-loop
result = []

for value in ords_prods_merged["orders_day_of_week"]:
  if value in [0,1]:
    result.append("Busiest days")
  elif value in [3, 4]:
    result.append("Slowest days")
  else:
    result.append("Regularly busy")

# Check output
result

# Incorporate results in df as a new column
ords_prods_merged['busiest_days'] = result

# 3. Check the values of this new column for accuracy. Note any observations in markdown format
ords_prods_merged['busiest_days'].value_counts(dropna = False)

# 4. Create a new column called “busiest_period_of_day” containing the labels “Most orders,” “Average orders,” and “Fewest orders”
# Find the frequency of the time column
ords_prods_merged['order_hour_of_day'].value_counts(dropna = False)

# Create a list using a for-loop
result = []

for value in ords_prods_merged["order_hour_of_day"]:
  if value in [10, 11]:
    result.append("Most orders")
  elif value in [3, 4]:
    result.append("Fewest orders")
  else:
    result.append("Average orders")

# Check output
result

# Incorporate results in df as a new column
ords_prods_merged['busiest_period_of_day'] = result

# 5. Print the frequency for this new column
ords_prods_merged['busiest_period_of_day'].value_counts(dropna = False)

# 7. Export dataframe as .pkl
ords_prods_merged.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_merged_new.pkl'))
