==> 2. Scripts/9. Crosstabs.ipynb <==
# Import libraries
import pandas as pd
import numpy as np
import os

path = r'C:\Users\opetr\Documents\Instacart Basket Analysis'

df = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'instacart_final.pkl'))

crosstab = pd.crosstab(df['days_since_prior_order'], df['order_number'], dropna = False)

crosstab.to_clipboard()
