==> 2. Scripts/10. Bussiness Questions 1.ipynb <==
# 01. Import libraries

#Import libraries
import pandas as pd
import numpy as np
import os

# 02. Import data

#Create path
path = r'C:\Users\opetr\Documents\Instacart Basket Analysis'

#3. Import the orders_products_combined dataframe from the pickle file you just saved
df_ords_prods_comb = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_combined.pkl'))

#Check output
df_ords_prods_comb

#4. Check the shape of the imported dataframe
df_ords_prods_comb.shape

#5. Combine the orders_products_combined dataframe with your products data set
##Import products data set
df_prods_cleaned = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'products_checked.csv'), index_col = False)

#Check output
df_prods_cleaned.head()

df_prods_cleaned.shape

#Drop '_merge' column from df_ords_prods_comb df
df_ords_prods_comb = df_ords_prods_comb.drop(columns = ['_merge'])

#Check output
df_ords_prods_comb

#Merge the two df
df_ords_prods_merged = df_ords_prods_comb.merge(df_prods_cleaned, on = 'product_id', indicator = True)

#6. Confirm the results of the merge using the merge flag
df_ords_prods_merged.head()

df_ords_prods_merged['_merge'].value_counts()

df_ords_prods_merged.shape

#7. Export the newly created dataframe as orders_products_merged in a suitable format
df_ords_prods_merged.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_merged.pkl'))
