==> 2. Scripts/12. Bussiness Questions 3.ipynb <==
# 01. Import libraries

# Import libraries
import pandas as pd
import numpy as np
import os

# 02. Import data

# Create path
path = r'C:\Users\opetr\Documents\Instacart Basket Analysis'

# Import the most recent dataset
df_cust = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'customers.csv'))

# 03. Explore data

df_cust.head()

df_cust.tail()

df_cust.info()

df_cust.dtypes

df_cust.describe()

df_cust.shape

# 04. Wrangle data

# Rename columns
df_cust.rename(columns= {'First Name' : 'first_name', 'Surnam' : 'last_name', 'Gender' : 'gender', 'STATE' : 'state', 'Age' : 'age', 'n_dependants': 'nr_of_dependents', 'fam_status' : 'marital_status'}, inplace = True)

# Check changes
df_cust.head()

#### Other types of data wrangling would be dropping a column, changing the data type, or transposing. None of these procedures are necessary for this dataframe

# 05. Check data quality and consistency

# Check if df contains mixed-type columns
for col in df_cust.columns.tolist():
  weird = (df_cust[[col]].applymap(type) != df_cust[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_cust[weird]) > 0:
    print (col)

# Change data type to string
df_cust['first_name'] = df_cust['first_name'].astype('str')

# Re-check if df contains mixed-type columns
for col in df_cust.columns.tolist():
  weird = (df_cust[[col]].applymap(type) != df_cust[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_cust[weird]) > 0:
    print (col)

# Check for missing values
df_cust.isnull().sum()

# Check for duplicates
df_cust[df_cust.duplicated()]

# Re-check shape
df_cust.shape

# 06. Combine df_cust with ords_prods_merged_grouped

# Import orders_products_merged_grouped data
ords_prods_merged_grouped = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_merged_grouped.pkl'))

# Merge the 2 dataframes
ic_final = ords_prods_merged_grouped.merge(df_cust, on = 'user_id')

# Check head
ic_final.head()

# Check shape
ic_final.shape

# Drop columns to decrese df size
ic_final.drop(['Unnamed: 0_x', 'add_to_cart_order', 'reordered', 'Unnamed: 0_y', 'busiest_day', 'busiest_period_of_day', 'median_order_frequency','first_name', 'last_name','date_joined'], axis = 1, inplace = True) 

# Check output
ic_final.columns

# Check new shape
ic_final.shape

# 07. Export datafrane as a pickle file

# Export new df
ic_final.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'instacart_final.pkl'))
